src/backend/fastapi-server/test-requirements.txt
uvicorn[standard]
fastapi
pydantic
requests
pymongo[srv]
bcrypt
pyjwt
httpx
fastapi.testclient
pytest